// apps/chat/utils/types.ts
export type CreateUserDetails = {
  username: string;
  password: string;
  firstName: string;
  lastName: string;
  peer?: any;
};